pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tinyllm"
version = "0.1.0"
description = "A neural network of LLMs - self-improving multi-agent system using small language models"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "TinyLLM Contributors" }
]
keywords = [
    "llm",
    "ai",
    "machine-learning",
    "neural-network",
    "ollama",
    "langchain",
    "langgraph",
    "multi-agent",
    "self-improving",
    "local-llm",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "types-PyYAML>=6.0.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.25.0",
]

[project.scripts]
tinyllm = "tinyllm.cli:app"

[project.urls]
Homepage = "https://github.com/ndjstn/tinyllm"
Documentation = "https://github.com/ndjstn/tinyllm#readme"
Repository = "https://github.com/ndjstn/tinyllm"
Issues = "https://github.com/ndjstn/tinyllm/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/tinyllm"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
]

[tool.ruff.lint.isort]
known-first-party = ["tinyllm"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "langgraph.*",
    "langchain_core.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/tinyllm"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

[dependency-groups]
dev = [
    "matplotlib>=3.10.8",
]
